+++ src/model/Rental.java
@@ -10,7 +10,7 @@ public class Rental
 { 
 	private int id;
 	private Date rentDate;
-	private int rentDuration;
+	private double rentDuration;
 	private Double total;
 	private int userID;
 	
@@ -23,7 +23,7 @@ public class Rental
 		
 		ps.setInt(1, id);
 		ps.setDate(2, rentDate);
-		ps.setInt(3, rentDuration);
+		ps.setDouble(3, rentDuration);
 		ps.setDouble(4, total);
 		ps.setInt(5, userID);
 		
@@ -46,11 +46,11 @@ public class Rental
 		this.rentDate = rentDate;
 	}
 
-	public int getRentDuration() {
+	public double getRentDuration() {
 		return rentDuration;
 	}
 
-	public void setRentDuration(int rentDuration) {
+	public void setRentDuration(double rentDuration) {
 		this.rentDuration = rentDuration;
 	}
